full serialization --part 3
