+++ Slider/Assets/Scripts/Localization/Phases.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 45de38b0c2644302bb92a162ede47caf
+timeCreated: 1735758365
