+++ Slider/Assets/Scripts/Localization/Phases/Localization.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 0346544fd3fa43918facdfd55341ecf5
+timeCreated: 1735758389
