+++ Slider/Assets/Scripts/Localization/Phases/Selection.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: fe968ed5be4047ffb5086573bb45cd72
+timeCreated: 1735758377
