+++ Slider/Assets/Scripts/Localization/Phases/Serialization.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 9a1a540ad8df403f969b021329129143
+timeCreated: 1735758296
