+++ Slider/Assets/Scripts/Localization/Utils.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a61d53b127ec40e6afe92d891bd23971
+timeCreated: 1735750496
